Move dqn model to models/
